+++ lib/github_copier.rb
@@ -3,5 +3,4 @@ require 'github_copier/opt_parser'
 require 'github_copier/github_copier'
 require 'github_api'
 require 'filename_cleaner'
-require 'ap'
 include GithubCopier