+++ lib/github_copier/opt_parser.rb
@@ -5,14 +5,13 @@ module GithubCopier
     # Return a structure describing the options.
     def self.parse(args)
       # The options specified on the command line will be collected in *options*.
-      # We set default values here.
       options = OpenStruct.new
 
-      # The mandatory arguments
+      # Set the sensible default for the options explicitly
       options.base_dir = "."
-
-      # The option arguments
-      options.verbose = false
+      options.all_repos = false
+      options.group_by_user = false
+      options.clone_repos = false
 
       # The parser
       opt_parser = OptionParser.new do |opts|
@@ -23,44 +22,50 @@ module GithubCopier
 
         # Mandatory argument
         opts.on("-b", "--base-dir BASE_DIR",
-                "Output directory where the repository will be cloned to") do |dir|
+                "where BASE_DIR is the directory where the repositories will be cloned to (mandatory)",
+                "If not specified, current directory will be used") do |dir|
           options.base_dir = dir
         end
 
         opts.on("-u", "--user USER",
-                "The github USER that will be cloned from") do |user|
-          # By default the id is assumed to be of a normal user not as organization
+                "The Github USER that will be cloned from (mandatory)") do |user|
           options.user = user
         end
 
         opts.on("-o", "--org [ORG]",
-                "The Github's organization name to be used if specified",
-                "(where ORG is the organization that the user belongs to)") do |org|
+                "The Github's organization name to be used if specified (optional)",
+                "where ORG is the organization that the user belongs to") do |org|
           options.org = org
         end
 
         opts.on("-t", "--oauth-token [OAUTH_TOKEN]",
-                "The Github's oauth_token for authentication (required to list/clone private repositories)",
-                "(where OAUTH_TOKEN is from the user's Github setting)") do |token|
+                "The Github's oauth_token for authentication (optional - only required to list/clone private repositories)",
+                "where OAUTH_TOKEN is from the user's Github setting") do |token|
           options.oauth_token = token
         end
 
         opts.on("-l", "--language [LANG]",
-                "Clone only project of type LANG",
-                "(where LANG is main language as shown on Github)") do |lang|
+                "Clone only project of type LANG (optional)",
+                "where LANG is main language as shown on Github") do |lang|
           options.language = lang
         end
 
         # Boolean switch.
         opts.on("-a", "--[no-]all-repos",
                 "All repository only (optional)",
-                "(default to original/non-fork repositories only)") do |a|
+                "default to original/non-forked repositories only") do |a|
           options.all_repos = a
+        end
+
+        opts.on("-g", "--[no-]group-by-user",
+                "Group the output by {BASE_DIR}/{USER}/{LANG}",
+                "default to {BASE_DIR}/{LANG}/{USER}") do |gbu|
+          options.group_by_user = gbu
         end
 
         opts.on( "-c", "--[no-]clone",
                 "Clone the repositories to the path specified (optional)",
-                "(default to --no-clone e.g. dry-run only)") do |c|
+                "default to --no-clone e.g. dry-run only") do |c|
           options.clone_repos = c
         end
 
@@ -68,9 +73,17 @@ module GithubCopier
         opts.separator "Common options:"
 
         # No argument, shows at tail.  This will print an options summary.
-        # Try it and see!
         opts.on_tail("-h", "--help", "Show this message") do
           puts opts
+          puts ""
+          puts "Example Usage:"
+          puts ""
+          puts "a) List the 'JavaScript' repositories for a given user (dry-run)"
+          puts "github_copier -b ~/Desktop/projects -u awesome_user -l JavaScript"
+          puts ""
+          puts "b) Clone the 'JavaScript' repositories for a given user (note: --clone or -c option)"
+          puts "github_copier -b ~/Desktop/projects -u awesome_user -l JavaScript -c"
+          puts ""
           exit
         end
       end