+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 #### 0.1.0
 
 - Initial release
+
+#### 0.2.0
+- Add option to group output by language by default
+- Show list of repositories in the dry-run mode (--no-clone) option
+- Improve the gemspec to add more context for the user
+- Add some example to default option when the user type no argument